# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_financeapp_backend_internal_domain.Category:
    properties:
      color:
        type: string
      created_at:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_financeapp_backend_internal_domain.Expense:
    properties:
      amount:
        type: number
      category:
        $ref: '#/definitions/github_com_financeapp_backend_internal_domain.Category'
      category_id:
        type: string
      created_at:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      import_id:
        type: string
      tags:
        type: string
      type:
        $ref: '#/definitions/github_com_financeapp_backend_internal_domain.TransactionType'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_financeapp_backend_internal_domain.Import:
    properties:
      created_at:
        type: string
      error_log:
        type: string
      errors:
        type: integer
      file_name:
        type: string
      file_type:
        type: string
      id:
        type: string
      imported:
        type: integer
      imported_at:
        type: string
      status:
        type: string
      total_rows:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_financeapp_backend_internal_domain.TransactionType:
    enum:
    - expense
    - income
    type: string
    x-enum-varnames:
    - TransactionTypeExpense
    - TransactionTypeIncome
  github_com_financeapp_backend_internal_domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_financeapp_backend_pkg_response.Body:
    properties:
      message:
        type: string
    type: object
  internal_auth.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/github_com_financeapp_backend_internal_domain.User'
    type: object
  internal_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_auth.RegisterRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_category.UpsertRequest:
    properties:
      color:
        type: string
      icon:
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - name
    type: object
  internal_expense.CategorySummary:
    properties:
      category_id:
        type: string
      category_name:
        type: string
      count:
        type: integer
      total:
        type: number
    type: object
  internal_expense.CreateRequest:
    properties:
      amount:
        type: number
      category_id:
        type: string
      date:
        type: string
      description:
        maxLength: 255
        minLength: 1
        type: string
      tags:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/github_com_financeapp_backend_internal_domain.TransactionType'
        enum:
        - expense
        - income
    required:
    - amount
    - date
    - description
    - type
    type: object
  internal_expense.MonthlySummary:
    properties:
      month:
        type: string
      total_earned:
        type: number
      total_spent:
        type: number
    type: object
  internal_expense.PaginatedResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_financeapp_backend_internal_domain.Expense'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  internal_expense.UpdateRequest:
    properties:
      amount:
        type: number
      category_id:
        type: string
      date:
        type: string
      description:
        maxLength: 255
        minLength: 1
        type: string
      tags:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/github_com_financeapp_backend_internal_domain.TransactionType'
        enum:
        - expense
        - income
    type: object
info:
  contact: {}
  description: 'Personal finance tracker: transactions, categories and bank statement
    imports.'
  title: FinanceApp API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Log out
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      summary: Register a new user
      tags:
      - auth
  /categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_financeapp_backend_internal_domain.Category'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Category payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_category.UpsertRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_internal_domain.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Create a category
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Transactions in the category are kept, with no category.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - categories
    put:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_category.UpsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_internal_domain.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - categories
  /expenses:
    get:
      description: Pass category_id=none to list only the transactions with no category.
      parameters:
      - description: Category UUID, or \
        in: query
        name: category_id
        type: string
      - description: Transaction type
        enum:
        - expense
        - income
        in: query
        name: type
        type: string
      - description: Start of the range (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End of the range (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Rows per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_expense.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: List transactions
      tags:
      - expenses
    post:
      consumes:
      - application/json
      parameters:
      - description: Transaction payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_expense.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_internal_domain.Expense'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Create a transaction
      tags:
      - expenses
  /expenses/{id}:
    delete:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Delete a transaction
      tags:
      - expenses
    get:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_internal_domain.Expense'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Fetch one transaction
      tags:
      - expenses
    put:
      consumes:
      - application/json
      description: Only the fields present in the payload are changed.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to change
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_expense.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_internal_domain.Expense'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Update a transaction
      tags:
      - expenses
  /imports:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_financeapp_backend_internal_domain.Import'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: List past imports
      tags:
      - imports
    post:
      consumes:
      - multipart/form-data
      description: Accepts CSV and OFX/QFX. Returns the import job with its row counts.
      parameters:
      - description: Statement file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_internal_domain.Import'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Import a bank statement
      tags:
      - imports
  /imports/{id}:
    delete:
      description: Deletes the transactions created by the import, then the job itself.
      parameters:
      - description: Import ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Revert an import
      tags:
      - imports
  /reports/categories:
    get:
      description: Defaults to the current month when no range is given.
      parameters:
      - description: Start of the range (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End of the range (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_expense.CategorySummary'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Totals per category
      tags:
      - reports
  /reports/monthly:
    get:
      description: Always returns the last 12 months; the months parameter is currently
        ignored.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_expense.MonthlySummary'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_financeapp_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Totals per month
      tags:
      - reports
securityDefinitions:
  BearerAuth:
    description: 'Paste the token returned by /auth/login as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"